Loose Row.get()
Previously, the Row interface only had a method of get(Object identifier,
Class type) that allowed a user to specify the type that they expected to be
returned from the call.  While the user could specify Object.class in order to
allow a loose typing, this was not clear in the documentation or API.  This
change updates the documentation to mention allowed variance and adds a
default method which encapsulates this behavior.

[resolves #4]