Remove executeReturningGeneratedKeys()
Previously, the Statement SPI contained a method that would return generated
keys in INSERT statements.  Given that this was a very specific use-case,
required a lot of insight by the driver into the SQL that was written, and was
supported in vastly different ways by different databases, this wasn't a great
method to have.  This change removes it completely, pushing the burden of
retrieving the generated keys to SQL authors and clients.

[resolves #17]