ConnectionFactoryProvider Improvements
As the first implementation of ConnectionFactoryProvider starts, a couple of
obvious improvements jumped out.  First, the methods on ConnectionFactories
should be static (class maintains to state).  Second many implementations will
need to check for the existence of Options in supports() without examining the
value.  An additional hasOption() method simplifies this behavior.

Signed-off-by: Ben Hale <bhale@pivotal.io>