Remove Recursive Generics
Previously, a change added recursive generics to the API in order to
facilitate direct use of the implementations.  After review, it was pointed
out that using these recursive generics makes it very difficult for clients to
consume the API.  Combined with a prototype that illustrated that we don't
internally make use of the APIs in a way that is helped by the recursive APIs,
this change reverts the addition of those generics.

[resolves #24]

Signed-off-by: Ben Hale <bhale@pivotal.io>