Extend ColumnMetadata
ColumnMetadata now exposes additional metadata information on a best-effort
basis. Drivers are free to provide metadata if it is available as part of a
result response or statically available.

Drivers also can decide which methods to implement and they can expose a
native metadata object if they can expose additional metadata.

[resolves #42]

Signed-off-by: Ben Hale <bhale@pivotal.io>