BLOB/CLOB Test Improvements
Previously the BLOB and CLOB tests assumed BLOB and CLOB were supported data
types in all databases.  This is not true, specifically for PostgreSQL.  This
change makes those types configurable and makes some general improvements to
the tests themselves to use JDBC as an independent check on the R2DBC
implementations.

In addition MockBlob and MockClob types were introduced to help ensure that
discard() was called.

[#41]

Signed-off-by: Ben Hale <bhale@pivotal.io>