IsolationLevel as Extensible Constants
Previously, IsolationLevel was an enumeration with four fixed values.
However, databases support more than the four SQL standard isolation levels
with proprietary extensions.  The previous enumeration strategy was
insufficient for this kind of extensibility.

This change modifies the enumeration to be a class, based on the previously
created ConstantPool, with constants for the four standard isolation levels,
but which can be extended as necessary by databases for proprietary
extensions.

[resolves #82]

Signed-off-by: Ben Hale <bhale@pivotal.io>