[#27] Offer result type agnostic mapping to Result
Clients may not know in advance whether a Result contains update counts or row data. The current SPI design does not allow for checking the Result for its contents, nor for trying out both possibilities and reverting to the other if one fails.

This suggestion offers a solution to this problem:

- A new Mapping<T> type is added containing functions for the individual mappings. This type is important for forwards compatibility, such that additional types of results (e.g. out parameters or signals, exceptions, warnings) can be added later
- The new Mapping<T> allows for mapping between 1 element (update count, row) to 0-N elements via a Publisher<T>. This is useful to allow for skipping values directly on the individual Result level
- A new map(Mapping<T>) method is added as a breaking change
- Default implementations delegating to the new methods are added for the existing methods

Signed-off-by: Lukas Eder <lukas.eder@gmail.com>